+++ packages/ove-core/src/client/utils/utils.js
@@ -141,9 +141,6 @@ function OVEUtils () {
         log.info('Initializing controller with state:', state);
         //-- The default state URL is used here. --//
         $.ajax({ url: 'states/' + state, dataType: 'json' }).done(initMethod).catch(log.error);
-
-        //-- BACKWARDS-COMPATIBILITY: For <= v0.3.3 --//
-        $.ajax({ url: 'state/' + state, dataType: 'json' }).done(initMethod).catch(log.error);
     };
 
     this.initControl = function (defaultState, initMethod) {