+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.4.0"></a>
+# [0.4.0](https://github.com/ove/ove/compare/v0.3.3...v0.4.0) (2019-04-08)
+
+### Features
+
+**feat:** Section-specific flush operation.
+**feat:** Support for validating application state on a per-application basis.
+
+### Fixes
+
+**fix:** Standardising OVE APIs.
+**fix:** Version upgrades to all dependencies.
+
 <a name="0.3.3"></a>
 # [0.3.3](https://github.com/ove/ove/compare/v0.3.2...v0.3.3) (2019-02-27)